rename d.ts to just .ts due to import error
