+++ packages/jellyfish-crypto/src/bip66.ts
